fix: idempotent apply, surfaced conflict warnings, fractional channel duration
Three bugs found during live multi-block testing against Tunarr 1.3.13.

CRITICAL: apply was non-idempotent. The 24h apply window and default 6h
cron interval mean the same future occurrence falls inside several
consecutive applies' windows; every one re-ran PlanBlock from scratch,
advancing series cursors and writing schedule_history for content that
was never going to air differently. A Saturday-night block observed E01
then E04 after two more applies; a show's series_state drifted to S2E3
in a day of test applies. Conflict-dropped occurrences had it worse --
PlanBlock ran and advanced state for every planned occurrence before
conflict resolution discarded the losers.

Fixed with per-(block_name, occurrence_start) idempotence:
GenerateForTimeRange now runs in three phases (build occurrence shells,
resolve conflicts on the shells, plan content only for survivors) so a
dropped occurrence never reaches PlanBlock. Filter blocks replay an
already-committed occurrence verbatim (no cursor to re-derive). Series
blocks needed more, since users edit a block's series order/membership/
episodes_per_block/duration before an occurrence airs and expect that to
take effect: a series occurrence's per-show cursor snapshot is captured
once, immutably, the first time it's planned (new
series_occurrence_snapshots table); a still-future occurrence re-derives
every apply from that fixed snapshot plus the CURRENT spec (idempotent
when unchanged, responsive to edits); once aired, it's frozen and replayed
verbatim forever. schedule_history gained occurrence_start/sequence/
duration_ms/title/type columns to support the per-occurrence lookup.
Also fixed a related RunCount inflation bug this surfaced
(markSeriesCompleted re-ran its bookkeeping on every re-examination of an
already-completed series within one wide window, not just once).

MAJOR: conflict resolution was invisible to API callers -- only a
server-side INFO log, nothing on the /generate or /apply response.
PlanResult gained an optional `warnings` array (api/openapi.yaml, both
codegens regenerated) listing each dropped occurrence's block name,
time, and blocker. The Schedule page now renders a warning panel after
preview/apply when any are present.

MINOR: tunarr.Channel.Duration was int64, which failed to decode a live
fractional-ms response (691200000.9999), and internal/httpclient's retry
condition fires on ANY error regardless of status -- so every such
decode failure cost up to 3 extra live requests for a response shape
that could never decode differently. Widened to float64, matching
Program.Duration's existing convention; the field is otherwise unused.

Regression tests: two consecutive Runner.Run(Apply:true) calls over the
same window produce byte-identical plans and unchanged series state; a
conflict-dropped series occurrence never creates a series_state row; the
exact reorder-before-airing scenario (commit [A,B,C], reorder to
[C,A,B], re-apply -- same occurrence, new order, same episodes, same
state); warnings round-trip through the API; fractional-duration decode
with a request-count assertion proving no retry loop.

+++ internal/external/tunarr/models.go
@@ -15,8 +15,23 @@ type Channel struct {
 	Name       string       `json:"name" validate:"required"`
 	Icon       *ChannelIcon `json:"icon,omitempty"`
 	GroupTitle string       `json:"groupTitle"`
-	Duration   int64        `json:"duration"`
-	Stealth    bool         `json:"stealth"`
+	// Duration is float64, not int64: live-verified this session that a
+	// real Tunarr instance's channel.duration (server-computed as the sum
+	// of its lineup items' own -- also float -- durations, see
+	// LineupRepository.ts's updateChannel) can come back with a fractional
+	// millisecond component (e.g. 691200000.9999), matching
+	// tunarr.Program.Duration's existing "float from Tunarr API"
+	// convention. An int64 field here used to make GetChannels's response
+	// decode fail on any such channel -- which, worse, is not a one-shot
+	// failure: internal/httpclient.Client's resty retry condition fires on
+	// any non-nil error regardless of HTTP status (see
+	// httpclient.DefaultConfig's AddRetryCondition), so a decode error
+	// alone triggered up to MaxRetries (3) extra live requests against
+	// Tunarr for a response shape that could never decode differently.
+	// This field is otherwise unused by any caller in this repo (grepped),
+	// so widening its type has no other effect.
+	Duration float64 `json:"duration"`
+	Stealth  bool    `json:"stealth"`
 }
 
 // Genre represents a genre with metadata.
